requirements.txt
pydantic>=2.6,<3.0
reportlab>=4.2,<5.0
python-dotenv>=1.0,<2.0
Flask>=3.0,<4.0

# RL agent
gymnasium>=0.29,<1.0
stable-baselines3>=2.3,<3.0
transformers>=4.40,<5.0
torch>=2.1
tqdm>=4.0
openenv-core
fastapi
uvicorn
httpx
openai
